+++ crates/rover-schema/src/describe/type_detail/fields/fields_detail.rs
@@ -1,0 +1,17 @@
+use super::field_info::FieldInfo;
+
+#[derive(Debug, Clone, serde::Serialize)]
+pub struct FieldsDetail {
+    fields: Vec<FieldInfo>,
+    pub field_count: usize,
+}
+
+impl FieldsDetail {
+    pub fn new(fields: Vec<FieldInfo>, field_count: usize) -> Self {
+        Self { fields, field_count }
+    }
+
+    pub fn fields(&self) -> &[FieldInfo] {
+        &self.fields
+    }
+}
