+++ crates/rover-schema/src/describe/type_detail/fields/mod.rs
@@ -1,0 +1,19 @@
+mod arg_info;
+mod enum_value_info;
+mod expanded_type;
+mod extended_fields_detail;
+mod field_detail;
+mod field_info;
+mod field_summary;
+mod fields_detail;
+mod type_kind;
+
+pub use arg_info::ArgInfo;
+pub use enum_value_info::EnumValueInfo;
+pub use expanded_type::ExpandedType;
+pub use extended_fields_detail::ExtendedFieldsDetail;
+pub use field_detail::FieldDetail;
+pub use field_info::FieldInfo;
+pub use field_summary::FieldSummary;
+pub use fields_detail::FieldsDetail;
+pub use type_kind::TypeKind;
